First Commit
Just a quick hello world to gain a bit more insight into python syntax
and loops. Frame Decomposition takes a video and seperates the frames
into jpegs. Then I've attempted to compare the difference between the
frames with a MSE. This difference will be saved to an array and used to
select frames for the new time lapse.