SSIM Fully Implemented
SSIM is now fully implemented. Now writing a method to work with a
moving window for frame selection. Initial results show videos made with
SSIM and MSE are not a massive improvement relative to the source.
However, this may be due to the relatively limited frames in the small
videos I've chosen. Additional tools like colour correction could also
help with this lack of smoothnesss.