More progress towards plotting
Switched froma  list comprehension to a numpy based 2D array for storing
our error values. This array is also saved to a text file now. The array
is initialised with all 0 values. The next step is to begin to create
our actual plot of the array and the ability to plot along additional
lines