Updated Error Plotting Now with Working Heat Map
Heat map component now works. Ultimately, the function should include
the similarity metric used as an arguement so that we can name the plots
and outputs uniquely. The plot should probably be rotated and scaled a
bit better but this is a good start. Saving the plot does not work at
the moment.