More progress
Now the DP algo seems to work. I'm saving the output frame list as well
and have created a new notebook to take this string and then create a
video using the list. Still need to finish the section of Path
Calculation where we calculate the final frame via a 2 dimensional
argmin.