Minor Fixes
Added a new method for cleaning timelapse footage. This is a bit more
straightforward and makes use of the filename for the timestamp.
Unfortunately filenames do not seem to be 100% accurate for when the
image was taken but thus far this is the best method. Also, small erors
fixed in path finding algo. More debugging to come